Feature/117#image

+++ src/main/java/store/novabook/front/api/book/dto/response/GetBookResponse.java
@@ -22,7 +22,7 @@ public record GetBookResponse(
 	List<String> tags,
 	List<String> categories,
 	int likes,
-	int score,
+	Double score,
 	String image,
 	Long bookStatusId
 ) {
